Stringify all the things
Also enforces stringification of all things sent to sidekiq. This means that stuff will fail if symboly keys are sent rather than just throwing an error message. This is in order that we can be more confident that all the logs we want are being sent. This will become mandatory in sidekiq 7 (coming soon from a publishing platform team near you)