Add content_ids to topic_groups definitions
This effectively removes the deprecated flag.

The content_ids property was added to enable support of service_manual_groups
as it was decided that the should use the existing topic groups definition
according to https://github.com/alphagov/govuk-content-schemas/pull/203
The schema was then split into "topic_groups" and
“service_manual_topic_groups" in
https://github.com/alphagov/govuk-content-schemas/commit/7bad6a241d3b3e178b66df9e81fd9cc8edc54e95
and the properties marked deprecated for later removal.

This also updates the description in line with the recently unified topic
terminology.