Manually set the path to content schemas
The new version of the govuk_schemas gem allows for the path to content schemas to be manually set, rather than using an environment variable (which happens for the deployed version of the app) or using a fallback (which happens for local development). This change means that this path will be used for both the running app and local development.

Also removed a message from the spec helper since it seems less relevant now that schemas are in publishing api.