Remove deprecated `description` from topic_groups content schema
This property was added to enable support of service_manual_groups
as it was decided that the should use the existing topic groups definition
according to https://github.com/alphagov/govuk-content-schemas/pull/203
The schema was then split into "topic_groups" and
“service_manual_topic_groups”<U+2028>in
https://github.com/alphagov/govuk-content-schemas/commit/7bad6a241d3b3e178b66df9e81fd9cc8edc54e95
and the properties marked deprecated "for later removal".
About time we get rid of it.