Remove symlink for allowed document types
This was put in temporarily while switching over to publishing api for content schemas, as content-publisher used it in test and was looking for it under the same path as in govuk-content-schemas. This has now been fixed via a bump to the govuk_schemas gem, which uses the updated publishing api path for this file.