Update jenkins file to run schema tests when there are any changes to content schemas
It was previously only configured to pick up changes to the dist directory. This is probably fine since dependent apps only really care about the compiled schemas. However, changing it so that removing the symlink to allowed-document-types.yml can be properly tested. We may want to revisit this change and switch it back if it slows down builds significantly.