Add migration for fixing bad organisation state
In https://github.com/alphagov/whitehall/pull/7154, we made changes to
how Whitehall deals with routes. This had the unfortunate side effect of
breaking some routing that was presented to Publishing API.

This left some organisations with a state that their translation(s) were
unpublished and redirected back to the English page.

This adds a migration which restores these organisations pages back to
their previous state, with the translations correct.