Include hide_from_external_crawlers in finder expansion rules
Finders have a feature that can allow them to be hidden from external
crawlers. We want this behaviour to be also exhibited in documents that
have an explicit association with the finder (which is the case for
specialist finders) as we consider these a child of the finder that
should follow the same crawler visibility rules.