Include no_index in finder expanded links
Finders have a feature that can allow them to set no_index to external
crawlers. We want this behaviour to be also exhibited in documents that
have an explicit association with the finder (which is the case for
specialist finders) as we consider these a child of the finder that
should follow the same no_index visibility rules.