Proof of concept for using pact artifacts
We have a problem that GitHub Actions fail when Dependabot raises a pull
request. The reason this fails is because Dependabot pull requests do
not have access to secrets and thus cannot write to the pact broker.

I'm trying out a different approach for the pact test and making use of
artifacts [1] as a way to share data between jobs to avoid the use of
the pact broker and it's requirement for secrets.

This is something of an experiment as I've not been able to fully assert
that Dependabot PRs will be able to create artifacts, but I've not found
anything to suggest they wouldn't. So, I'm hoping to get this merged in
and test it against our open Dependabot requests for Publishing API. It
only tests against a branch of content-store: https://github.com/alphagov/content-store/pull/1046

Should this be a success I plan to do a little bit of extra work to
neaten things up and get the change to Content Store merged in. Then I
plan to take the same approach with GDS API Adapters, which also has
this problem (however it's less visible as Dependabot PRs are much less
frequent).

As an aside, it's driving me a bit crazy using the US spelling for
artifact and deciding whether to use the British one for non-noun
contexts or just be consistent throughout 🙀.

[1]: https://docs.github.com/en/actions/using-workflows/storing-workflow-data-as-artifacts