Clean up GitHub Action pact handling following proof of concept
This depends on: https://github.com/alphagov/content-store/pull/1046

In https://github.com/alphagov/publishing-api/pull/2276 we explored a
proof of concept for having Pact test work with GitHub Actions for
Dependabot PRs. This proof of concept was a success and this cleans up
the approach for prosperity.

I've removed the step to publish pact tests to the Pact Broker. I
realised this was of low value as the Content Store does not use the
`branch-main` [1] and instead relies on Jenkins CI. We can always
reintroduce this when we have a need for it, I just wanted to not leave
something in that didn't provide value.

The other change I've made is the deletion of the artifact after the
action run, since I can't think of a situation where we'd need this and
we can avoid this contributing to our storage quotas.

[1]: https://github.com/alphagov/content-store/blob/685d3d88e59f3066740bdc3ae177ca811f66d060/.github/workflows/ci.yml#L26-L30