Run GDS API Adapter pacts during CI
This app wasn't previously configured to run the pact tests via ci.yml
as other apps are and instead was using the push/pull_request events.
This makes this project consistent with other GOV.UK repos.

I've also named the workflow specifically GDS API Adapters as it's quite
confusing how this app is a provider to GDS API Adapters and a consumer
of Content Store.