Add required fields for historic appointments to link expansion
This adds expansion rules for the historic appointments document type (which is currently just for the past prime ministers index page). It adds fields for link expansion that will enable us to render the information we need on the page.

This enables us to each past prime minister's name, historical account href, and image details to render on the page.

We needed to define a recursive link for the historical account => person, since the image is contained in the person, which is a link in the historical account. So this ensures that the person link is also expanded, within the expanded historical account link.