Change dates in office on Historic Appointments to be an array
We are currently including a string for dates of historic appointments,
such as `2010 to 2015`. This won't allow us to cleanly group the
appointments by year in the frontend.

Therefore updating this to use an array, in the same way the Historic
Apppointments format already does.

Additionally, the field name has been updated to match the other format,
to remove inconsistency.