Fix magic class names for DB migrations.
7e66d46 broke database migrations whose class names contained the string
`Json`, because it changed the logic for magically generating the
camel-cased class name from the source filename.

Update the class names to match what Rails now expects them to be.

Perfect illustration of why such magic is a terrible idea, but Rails
makes it hard to avoid this sort of thing.