Add image using shared image definition to historic appointments
Using the shared definition for images will allow us to include alt text for images for recent past prime ministers without individual historic accounts, and also to standardise the way we handle these images on the frontend, since they will be presented in the same structure for past prime ministers with and without historical accounts.

Note that this is going to be done in a two-stage process - first adding the new image field (which isn't required), then updating the publishing application to populate this instead of the image_url (which again isn't required), then removing the redundant image_url. This has been done to ensure a smooth transition and means that we'll be able to get a passing test suite of dependent applications for this PR.