Add migration to remove path reservation for `/world`
There is a path reservation for `/world` reserved by Content Tagger. No
content is currently published to this path and there is no intention
for Content Tagger to use this path for anything.

Therefore deleting it so Whitehall is able to publish the world index
page at the same path.

Adding this as a migration so there is a record of when this was removed
(and why), and also to make it safer to run than executing commands on a
console.

The safe navigation operator has been used as local installations of
Publishing API may not contain this path in their database.