Update embassies_index schema
We are moving the rendering of gov.uk/world/embassies from Whitehall
to a rendering app (probably collections). To do this we first need to
publish a content item for the "embassies index" page that corresponds
to this schema.

On the page in question each world location can have either:

- a list of one or more local offices that British nationals can
contact for assistance (e.g. Australia has "Canberra", "Sydney"
etc.)

- a single remote office that British nationals can contact for
assistance (e.g. Afghanisation currently has a remote office in
Qatar)

- no offices that a British national can contact (in which case we
currently display "contact local authorities").

We've modelled those three cases in this schema using three defined
subschemas.

Note that we initially made the `world_locations` property required but
realised we can't do that because it causes the existing document we're
publishing to /world/embassies from Whitehall to fail the validation[1].
Once we've switched Whitehall to send the real data of the
/world/embassies page then we can make the `world_locations` property
required again.

[1]: https://github.com/alphagov/whitehall/blob/main/lib/publish_static_pages.rb

Co-authored-by: Chris Lowis <chris.lowis@digital.cabinet-office.gov.uk>