Add whip organisation to ministerial role details
We're in the process of moving the rendering of the Ministers page from
whitehall into collections.

At the moment, the ministers content item includes (via expansion) all
the person's roles. For most page sections, this is fine, as we show
all of a minister's current roles.

However, the Whips section only needs to show current whip roles, so we
need to get the role's whip organisation information into the content
item.

Co-authored-by: BeckaL <becka.lelew@digital.cabinet-office.gov.uk>