Expand ordered_parent_organisations for ministers
Collections is going to be responsible for rendering the Ministers index
page, which includes a Ministers by department section showing each
department's ministers and their relevant roles.

Whitehall publishes the departments and associated ministers, with their
correct order, but the roles are added later, in the link expansion
process. This lists all the roles, whether they are relevant for the
department being shown or not.

In order to show only a minister's relevant roles within the org being
shown, collections needs access to the organisations associated with
each role.