Regenerate worldwide_office schemas
To include the `marine_equipment_approved_recommendation` document type
in the worldwide_office schemas.

By running `rake build_schemas`.

The `marine_equipment_approved_recommendation` document type was added
in PR 2343[1] and the worldwide_office schema was added in PR 2361[2].
We're not sure why the tests didn't fail when the second of these was
merged into main.

[1]: https://github.com/alphagov/publishing-api/pull/2343
[2]: https://github.com/alphagov/publishing-api/pull/2361

Co-authored-by: Chris Lowis <chris.lowis@digital.cabinet-office.gov.uk>