Remove unpublishing document_type and schema
This schema is a remnant from the early days of Publishing API when an
unpublishing document type was used to remove content. However this was
phased out by V2 of the Publishing API in 2015 with a dedicated
unpublishing endpoint [1].

This schema no-longer serves a purpose and no-content is using it.

[1]: https://github.com/alphagov/publishing-api/blob/9f02226ce6827da8a3e23a6229f717c713bae1b2/docs/api.md#post-v2contentcontent_idunpublish