Remove news_article document type
This document_type seems to have been added by mistake as news_article
is only the name for a schema. The document types that match that schema
are: "press_release", "news_story", "government_response" and "world_news_story".

There are no current documents in the Publishing API that use this
document type.