Add examples for ministers index schema
This is an already edited version of the real content item. I've kept
enough items to have all the interesting test cases that we have
discovered in the process of rendering the Ministers index page in
collections.

I've added an example for a regular item (when there is no reshuffle
going on) and an example for when a reshuffle is ongoing.