Add multi level links Worldwide Organisations
We are currently retrieving the Worldwide Organisation that a Worldwide Office
belongs to through the `parent` link.

However, we now also need to be able to retrieve the Sponsored Organisations
and World Locations that are linked to the parent Worldwide Organisation.

In order to define multi level links, we need the link to be named (instead of
just being a `parent` link).

This adds the required named link to Worldwide Offices, and defines multi
level links so that links to Worldwide Organisations include Sponsored
Organisations and World Locations.