Add people with roles to worldwide org
Whitehall currently shows any people who are appointed to certain roles
in a worldwide org in the following way:
- image and details for the person in a primary role, if they exist
- details only for people in secondary or office roles.

In order to render them in government-frontend, we need to have the
people and their roles in the content item.

As these new links include role appointments in the chain, they need to
be exempted from requiring base_path.

The person in a primary role needs to have image fields.

Whitehall distinguishes "secondary" and "office" roles in the backend,
but not when displaying them.

I chose to keep this distinction in the content item, to allow for
future developments if the distinction ever becomes meaningful for
displaying.