Link main and home page offices for worlwide orgs
In order to keep the current behaviour from whitehall, we need the
worldwide organisation content item to have the information about which
office, if any, is the main office, and which offices to show on the
home page of the organisation.

Worldwide offices have their own content item, and some of the data
(such as the `public_url` of the office to be shown when the main office
has access and opening times) belong to the office rather than to its
contact. So we have linked the offices rather than skipping directly to
the contacts' `content_id`.