Add secondary corporate info pages to example
The secondary pages belong to a different organisation than the org used
to generate the bulk of the example. There is no one org that cumulates
all the cases we need to deal with, and I find it better to have a rich
example centrally than to rely on individual rendering apps to define
the content and test against it.