Rename method on Redirect Presenter
There are two ways of representing a redirect in publishing api - one with a state of published and one with a state on unpublished with an associated 'redirect unpublishing' record. Currently, the redirect presenter only handles the second case. This renames a method to make it clear which version it is being instantiated from, ahead of introducing a method to handle the other type of redirect in here.