Add optional first published at date to redirect presenter
We currently use the redirect presenter for content items that have been unpublished. We are also now going to use them for content items with a state of published and a document type of redirect, since it makes sense to treat both types of redirects consistently.

The redirect presenter is more minimal than the edition presenter that live redirect content items were previously being presented by. From looking at examples of the two types of redirect in content store, the only obvious field that is missing for content items presented by the redirect presenter is first published at. Although it's not strictly necessary since the page is not rendered anywhere, I'm adding it in in case it's useful for debugging at any point without having to look in the publishing api database.