Add current prime minister to example
The corresponding test in government-frontend is passing, because it is
comparing that `nil` equals `nil`. However, we are getting a deprecation
warning from minitest:

```
HowGovernmentWorksPresenterTest
DEPRECATED: Use assert_nil if expecting nil from
test/presenters/how_government_works_presenter_test.rb:14. This will
fail in Minitest 6.
```

Add a non-nil item so that the test is meaningful.