Correct document type on call for evidence examples
We have two fields on call for evidence one is called schema and one
is called document_type. We need document_type to reflect the state
of the call for evidence so it can be correctly rendered.

This corrects where the document_type has been put as the schema so we
could work on the code in another app.