Make `rendering_app` optional for `world_location` schema
World Locations are not rendered by any application, as they have no
base path and no routes.

The content item exists solely for use in link expansion.

We currently set the `rendering_app` to `whitehall-frontend`, which is
incorrect as these pages are not rendered.

Therefore making the field optional so we can remove the rendering app
completely.