Rollback version of sidekiq-unique-jobs
This update seems to have broken our ability to update some pages. We
have been having an issue when making changes to certain pages and logging
an error related to non-unique `lock_digest`. This is a thing that comes
from sidekiq and more specifically the sidekiq-unique-jobs gem. This
rolls back an update to that gem that seems to have happened around the
same time that these errors started occurring. Rolling back in the hope
that it fixes things!