Remove unused debug code
This was added in https://github.com/alphagov/publishing-api/pull/345 to
help debug content issues.

It's not clear how widely known / used this feature was, but it was
broken in https://github.com/alphagov/publishing-api/commit/b7712f3ba7630aa87db2be9b1ac5f35e298131ed
when publishing-api switched to being an API only rails app. The
controller now returns HTTP 204 (no content) responses, because API only
rails apps don't do views like full rails apps do.

A quick strawpoll suggested that the prevailing view was that we should
remove the dead code.

We may want to reimplement some of this once we've worked out a pattern
for hosting administrative information / tasks in publishing-api (e.g.
we need a better way to debug sidekiq jobs). There was certainly a lot
of enthusiasm for it in the original PR, so it feels a bit sad to lose
it. But there's no point keeping code around if noone can use it -
that's what the git history is for.