Allow external redirects to caa.co.uk
This allows users to unpublish documents and redirect it to URLs on the
Civil Aviation Authority (CAA) website.

Requested in Zendesk ticket:
https://govuk.zendesk.com/agent/tickets/4994883

And agreed in Slack thread:
https://gds.slack.com/archives/CACV3TACU/p1694690016148749

I originally added this to Whitehall in PR alphagov/whitehall#8290
before realising that the document to be unpublished & redirected is
actually published via Specialist Publisher, and therefore this is
actually where the list needs updating. 🙈