Resolve issue when unpublishing redirect content items
When a content item with document type of `redirect` was being
unpublished as `gone`, we were still presenting it to Content Store with
the redirect presenter (instead of the gone presenter). This meant that
the redirect remained in place and the gone response was never handled.

This was because the `content_store_presenter` method prioritised the
redirect over the unpublishing, which is not correct. The unpublishing
should take priority over the document's existing type.

Therefore updating to ensure we handle unpublishing first.

Also added a test to cover the scenario, which was failing before this
fix was implemented.