Disallow a conflicting draft editions
Currently, a publishing application can send a draft edition to
Publishing API, which has the same base path as a live edition but with
a different content ID.

This results in a valid draft edition being created in Publishing API.

However after this occurs, neither the original live edition with the
original content ID or the draft edition can have a new edition created,
as there will be a conflict in content IDs.

This updates the validation to disallow a draft edition with a different
content ID if there is already a published edition at that base path.