Check response not the raised exception
This test started failing in Rails 7.1, as the default value of
`config.action_dispatch.show_exceptions` was changed from `false` to
`:rescuable`. This means we now render a nice 404 response rather than
raising an exception, but need to update the test to match this.

See https://mattbrictson.com/blog/rails-71-features#1-finally-we-can-write-proper-integration-tests-for-errors
for some more information.