Move ContentStoreWriter to `lib`
In Rails 7.1, the `add_autoload_paths_to_load_path` config is set to false by
default.

This means that the `ContentStoreWriter` class is unavailable in the `Services`
initializer.

Moving this class to the `lib` directory makes it available again; the lib
directory is not affected by the `add_autoload_paths_to_load_path` flag, it is
added to $LOAD_PATH always. The other classes loaded by the `Services`
initializer also live there.