Spike: serve content-store requests directly from Publishing API.
Basic copy-and-paste of find_by_path, which now works on editions.
Add controller in new content-store namespace to prove concept
of serving content-store requests directly from publishing-api.

The new controller responds on /content-store/(live or draft)/(path)
and uses the DownstreamPayload to present the relevant edition as JSON.

It takes just under 0.5s to respond on a local laptop. We'd need to
speed that up for production use, with some combination of optimisations
and caching. That's out-of-scope for this spike, we've proved that it's
potentially possible and that's enough.

linting