Add an AB test schema, and allow links to it
Idea here is to publish AB test config as documents, which can then be
linked to from other documents (e.g. Guides, Answers etc.).

The frontend applications can then use the linked information to
determine whether / how to set up an AB test for a particular content
item.

The idea with placeholder is that the frontend will find and replace the
`{{ab_test_example}}` string with one of the `variants` configured in
the AB test.

We could also automate writing the AB test configuration to the Fastly
dictionaries by listening for these items on the publishing queue and
creating / updating / deleting dictionary items as necessary.