Add destination to redirect validation errors
It's usually not obvious when you're looking at one of these 422
unacceptable entity validation messages in Sentry or Kibana what the
payload was, so an error message that just says "internal redirects
cannot end with /" can be hard to diagnose.

Adding the destination in these cases should make it much more obvious
what the validator is upset about.

We already include the destination in error messages in the
reject_query_parameters_and_fragment method, so I don't see any reason
why it would be a problem to include them here too.