Example of a recursive CTE with Arel
This was the best I could manage...

It feels like there should be a nicer way to do the Arel.sql("t(n)")
bit, and the Arel::Nodes::Union.new(...) bit, and the
Arel::Nodes::Cte.new(...) bit. But I can't work out what it is.

Handling the nodes directly at least gives us the flexibility to work
with the bits of ActiveRecord that work well, without dropping down to
string processing.