Update WorldwideOrganisation schema
Whilst WorldwideOrganisation has a default news image, it does not form
part of the schema like other content types and so will not be exposed
as a possible link.

Here we add the default news image to the `details` property of the schema.

The property is not required as it is not required on the content
itself and so that we can manage the deployment.