Include default news image in linked entity
A WorldwideOrganisation includes a default news image, this is currently
added to the main document by Whithall when it publishes, which is
different to how Organisation handles the same behaviour.

We want to make the two entities implementation the same, our first step
on this path is to include the default_news_image in a
worldwide_organisation when included as a link in another content item.