Bump `google-protobuf` to 3.25.3
This was removed from the Gemfile.lock during [the Ruby 3.3
update](https://github.com/alphagov/publishing-api/commit/959d49ff0ea1601dfd431a03715aaf058e7c899d)
but not removed as an actual dependency, therefore any subsequent runs
of `bundle install` have resulted in the incompatible version of the gem
being reinstalled.

Therefore upgrading to a version of `google-protobuf` that is compatible
with Ruby 3.3.