Specify content stores when fetching editions
We have a the following relevant indices:
- index_editions_on_document_id
- index_editions_on_document_id_and_content_store
- index_editions_on_document_id_and_state

The editions table is very large because it includes superseded
editions. In my database dump, we have:

    select count(*), state, content_store from editions group by state,
content_store order by count(*) desc;

    |---------|-------------|---------------|
    | count   | state       | content_store |
    |---------|-------------|---------------|
    | 8807932 | superseded  | null          |
    | 781323  | published   | live          |
    | 132576  | unpublished | live          |
    | 100785  | draft       | draft         |
    | 7928    | unpublished | null          |
    |---------|-------------|---------------|

i.e. the majority of editions are in the superseded state / null content
store.

TODO - can we ignore the unpublished editions in the null content
store??

This query can be called with thousands of content IDs in the

    WHERE "documents"."content_id" IN ($content_ids)

clause, so the database has to do a nested loop to join the document ids
which correspond to these content ids with the edition document ids.
This mean potentially thousands of index scans - one for each document.

Using the index_editions_on_document_id_and_content_store means that
these index scans don't have to check all the editions in the null
content store, which makes things much faster.

You might think Postgres should use index_editions_on_document_id_and_state,
but it looks like the query often has a clause like:

    ("editions"."state" IN ('published') OR "editions"."state" = 'unpublished' AND "unpublishings"."type" = 'withdrawal')

... and the additional unpublishings.type reference means that it
doesn't like to use that index.

It looks like this could be fixed by putting an additional

    AND ("editions"."state" IN ('published', 'unpublished'))

to reassure postgres that it's fine to use that index. This might be
less risky than filtering on content store, but it might also be a bit
weird...