Be more specific about the "unpublish" message
This should use our new "substitute" payload. Unfortunately, in the case
of redirects the logic isn't quite right, and this test now fails
because the "unpublish" message has a "redirect" payload instead of a
"substitute" payload.

I think this would only be a problem when the substitution replaces a
redirect - other types of document would be fine. But it just so happens
the test case uses a redirect.