Rename SubstitutionHelper#clear! to be more explicit
I think it would be nice for SubstitutionHelper to also be responsible
for clearing things which _aren't_ of the same locale and base path, but
having one method called `clear!` would make this confusing. So let's
give the method a more descriptive name, and see where the night takes
us.