Remove UK Market Conformity Assessment Bodies finder
The finder documents have all been unpublished. The finder itself was expected to be unpublished.
That work was done in 2023, by using the unpublish rake task. Nonetheless, the code was not removed.

A zendesk ticket was raised again to remove the finder as it was publicly available, in 2024.
Running a rake task such as `publish_finders` would have caused the finder to be published again.

Current work is to rerun the rake task and remove all code.
These changes will be followed by removing the finder code from search-api and specialist-publisher.