Add withdrawal to unpublishing cases
content_presenter is already the fallback presenter, so all this does is
suppress warnings for "Unexpected unpublishing type withdrawal". We're
getting a few thousand logs of this kind per week[0].

I'm pretty sure that content_presenter is the correct choice here, as it
will end up as an EditionPresenter, which has a specific withdrawal_notice
method.

[0] - https://kibana.logit.io/s/13d1a0b1-f54f-407b-a4e5-f53ba653fac3/goto/23eae00a8aae5313779a07d129ac4f19?security_tenant=global