CSV Report for word usage
Occasionally, we're asked to search for certain lists of words to see if they appear on GOV.UK.

This could be for a number of reasons, but most recently it's been to determine whether there's any offensive language on GOV.UK.

This adds a rake task to publishing-api which, given a list of words, will check the details of all published editions for those words.

It batches editions into groups on 1,000 using find_each. Performance here is very sensitive to index use - it needs to be able to order editions by ID, so we want it to use an index which includes id, but we also want to avoid scanning the huge number of superseded editions. The `content_store: "live"` condition allows postgres to use `index_editions_on_id_and_content_store` - if we only filter on `state: "published"` the database has to use a much less favourable index (since there's no single index on `state`).

I'm doing this in publishing-api rather than content store mostly because I can, and partly because I happen not to have a content-store database clone locally.

Heavilly based on https://github.com/alphagov/content-store/pull/1113, which is what we used last time we did this (before the content-store to postgres migration)