Associate Document and Link through LinkSet
This allows for several quality of life improvements.

Given a document, we can now get its link set links with:

    document.link_set_links

Which is only a minor improvement over:

    document.link_set.links

But given that we hardly ever care about the link_set itself, this feels like a nicer interface.

More importantly, we can use this in joins, e.g. doing:

    Edition.live.where(...).joins(document: :link_set_links)

Instead of:

    Edition.live.where(...).with_document.joins <<-SQL
      INNER JOIN link_sets ON link_sets.content_id = documents.content_id
      INNER JOIN links ON links.link_set_id = link_sets.id
    SQL

 Also, given a link, we can now do source_document and target_document, which are nicer than having to look things up by their content_ids.

 Note that because we use the same Link model for both "edition links" and "link set links" source_document will be nil for edition links (just like edition is currently nil for link set links). target_document will always be set, as links always have a target_document.