Overhaul of link expansion docs
This commit makes the differences between link set links and edition links more explicit. It also explains why publishing applications might use one or the other of those categories of link and why each of them exists.

It also attempts to explain dependency resolution in a bit more detail.

This commit also includes a large number of edits for clarity.