Republish drafts of other translations on PUT of new translation
When a draft translation is added and another translation of the same
content ID exists, we need to send the existing translation(s)
downstream so their `available_translations` links can be updated to
include the new translation.

We're not currently doing this, which leads to a situation where the
link to the new translation never appears on the draft of the existing
translations.