Add political flag to case study schema
It has been decided that for the July 2024 election it should be possible to apply history mode to case studies.

History mode docs are at https://docs.publishing.service.gov.uk/repos/whitehall/history_mode.html

This requires us to send the political flag to Publishing API as part of the "details" hash for case studies.