Temporarily pin ruby version to 3.3.1
Deployments to integration are currently broken due to the `jsonnet` gem
failing to install.

We believe this is due to an issue with recent govuk-ruby-builder
images, and using 3.3.3 (the latest patch at time of writing) causes the
`jsonnet` gem to fail to install. This is currently being investigated
but I'm pinning Ruby back to 3.3.1 to unblock deployments.

Issue open here: https://github.com/alphagov/govuk-ruby-images/issues/96