Add a ContentEmbedService
This takes a string of HTML with embed code and attempts to
embed the latest edition of the content with a particular
type/content ID. I’ve kept the view deliberately simple for
now. Can add more detail later.

There’s also a bunch of TODOs for behaviour we’ll need later.