Add links to content when embedded content found
When we PUT a new edition to the endpoint, we scan the details
(currently just `body`) to find embed code, and append these to
the links array, ready to be expanded. This ensures we send this
content store for futureproofing, but also ensure that the content
we’re sending is recorded as a dependency of the embdedded content,
so when the embedded content is changed, our content is republished.